// packages/kit/src/views/Developer/pages/Gallery/Components/stories/QRWalletGallery/types.tsx
// eslint-disable-next-line @typescript-eslint/naming-convention
export enum KEYRING_ACCOUNT {
  standard = 'account.standard',
  ledger_live = 'account.ledger_live',
  ledger_legacy = 'account.ledger_legacy',
}
// eslint-disable-next-line @typescript-eslint/naming-convention
export enum KEYRING_MODE {
  hd = 'hd',
  pubkey = 'pubkey',
}